updated README.md by adding min/complie sdk versions description
This update is important because many projects are either large-scale, commercial, or were created some time ago. Upgrading the compileSdk version is not always quick or straightforward—it can trigger a chain reaction of dependency and compatibility issues. In such cases, it’s crucial to know the required SDK versions to choose the appropriate library version that aligns with local project constraints. Unfortunately, this information was missing from the changelogs of all major releases. I had to research the codebase myself to identify the exact point where the compileSdk was raised from 34 to 35.